0.05 - the birth of the worker
Workers are now implemented in a basic form.
The workers got 3 states (idle, moving and working) with a form of state
machine controlling what the worker does next. the worker is currently
able to detect when a job is aviable, move to it and execute the job
untill it is gone. the only job in game at the moment is fellTrees, the
fellTrees job changes the tile from forest to planes, more functunality
is yet to be added.

SelectedJobTextController is a bugFixing utility that displays what job
is currently selected for placing, also have a "Add worker" button since
there is no way to get workers in game yet and it's crucual that the
workers positios does not have any decimals (fix that later. potential
long play issue)